proof: remove the lemma-discovery enumerator, keep the committed-lemma consumer
The brute-force discovery enumerator (`aver proof --discover` / `--emit-laws`)
never earned its keep: it closes none of the hard unproven tasks directly, the
proof-corpus coverage metric never invokes it, and the agent-driven Method path
deliberately forbids it (proposed laws must close self-contained, not via
enumeration). Everything it "delivered" was actually the consumer's work — it
only guessed a helper lemma that a human or the Method writes by hand.

Cut the enumerator and everything exclusive to it: the term/equation
enumerator, the VM refutation filter, the structural-brick generator, the
born-as-Aver `--emit-laws` sidecar, the candidate ranking/report, and the
`--discover` / `--emit-laws` / `--emit-laws-to` flags plus their proof-command
plumbing. The discovery corpus test and the enumerator unit tests go with them.

Keep the consumer (`committed.rs` + the surface hash + the SimpOverLemmas hook
in `cmd_proof`): the piece that actually flips a law to `universal` by re-pinning
an earlier proven law (or any committed lemma file) as a simp rule. It has zero
dependency on the enumerator and is the substrate the Method feeds, so a normal
`aver proof` and the law-feeds-law / dependency-module / decomposed-task proofs
are unchanged (proof_spec stays green).

The full enumerator (including the unreleased `calculate.rs` Lemma-Calculation
seed) is preserved at tag `archive/lemma-discovery-enumerator` for revival.
