+++ docs/lean.md
@@ -136,7 +136,7 @@ The recognizer families, roughly in routing order:
 - **synthesized lemmas about your functions** — when a function matches a conservative shape gate (for example the canonical string-position scanner), the compiler synthesizes and kernel-proves a companion lemma about it and uses it in higher strategies such as the decimal render/parse roundtrip. When the gate does not match, nothing is emitted.
 - **fallback** — bounded evidence only: per-sample and `_checked_domain` theorems via `native_decide`, plus the universal theorem with an honest `sorry`. `when`-laws get a guarded-domain enumeration instead; their bounded statements carry an explicit statement class and are never credited as universal.
 
-When the recipes run out, the escalation path is more Aver, not Lean: split the hard law into helper laws — each one is a runnable test in milliseconds — and once proven, earlier laws in the same file become rewrite ammunition for the laws below them. The opt-in, experimental `aver proof --discover` mode additionally conjectures and kernel-proves helper lemmas automatically.
+When the recipes run out, the escalation path is more Aver, not Lean: split the hard law into helper laws — each one is a runnable test in milliseconds — and once proven, earlier laws in the same file become rewrite ammunition for the laws below them. A proven helper law committed in scope is picked up automatically by the law below it (the `SimpOverLemmas` feedback loop); the `the-method` agent loop can propose those helper laws for you.
 
 Termination is part of the same honesty story: structural recursion over your own types and recognized well-founded shapes (for example quicksort's mutual recursion) emit genuine total definitions; the remaining recursive shapes are emitted fuel-wrapped, with the fuel budget derived from a synthesized size measure of the call-site arguments.
