+++ src/main.rs
@@ -388,9 +388,6 @@ fn main() {
             minimize,
             gate,
             write_baseline,
-            discover,
-            emit_laws,
-            emit_laws_to,
         } => {
             commands::cmd_proof(
                 file,
@@ -407,9 +404,6 @@ fn main() {
                 *minimize,
                 gate.as_deref(),
                 write_baseline.as_deref(),
-                *discover,
-                *emit_laws,
-                emit_laws_to.as_deref(),
             );
         }
     }
